[components] Reduce amount of duplicated code for button variants (#464)
* Revert "[components] Fix "input inside button"-issue, solves upload button not working in Firefox"

This reverts commit c919f2598f65a45aa119e9de1c7f1f506c996b4f.

* [components] Reduce amount of duplicated code for buttons components
