+++ packages/@sanity/components/src/buttons/AnchorButton.js
@@ -1,95 +1,3 @@
-import PropTypes from 'prop-types'
-import React from 'react'
-import Ink from 'react-ink'
-import styles from 'part:@sanity/components/buttons/default-style'
-import Spinner from 'part:@sanity/components/loading/spinner'
+import createButtonLike from './createButtonLike'
 
-export default class AnchorButton extends React.Component {
-  static propTypes = {
-    kind: PropTypes.oneOf(['default', 'simple', 'secondary']),
-    color: PropTypes.oneOf(['primary', 'success', 'danger', 'white']),
-    onClick: PropTypes.func,
-    children: PropTypes.node,
-    inverted: PropTypes.bool,
-    icon: PropTypes.func,
-    loading: PropTypes.bool,
-    ripple: PropTypes.bool,
-    className: PropTypes.string,
-    disabled: PropTypes.bool,
-    tabIndex: PropTypes.number
-  }
-
-  static defaultProps = {
-    ripple: true,
-    icon: null,
-    onClick() {},
-    kind: 'default'
-  }
-
-  constructor(...args) {
-    super(...args)
-
-    this.handleClick = this.handleClick.bind(this)
-
-    this.state = {
-      recentlyHovered: false
-    }
-  }
-
-  handleClick(event) {
-    this.setState({
-      recentlyHovered: true
-    })
-    this.props.onClick(event)
-    clearTimeout(this.timeOut)
-    this.timeOut = setTimeout(() => {
-      this.setState({
-        recentlyHovered: true
-      })
-    }, 300)
-  }
-
-  componentWillUnmount() {
-    clearTimeout(this.timeOut)
-  }
-
-  render() {
-
-    const {kind, ripple, disabled, inverted, color, icon, loading, className, children, ...rest} = this.props
-
-    const Icon = icon
-
-    const style = [
-      inverted && styles.inverted,
-      kind ? styles[kind] : styles.default,
-      color && styles[`color__${color}`],
-      icon && styles.hasIcon,
-      className,
-      disabled && styles.disabled,
-      this.state.recentlyHovered ? styles.recentlyHovered : styles.notRecentlyHovered
-    ].filter(Boolean).join(' ')
-
-    return (
-      <a
-        {...rest}
-        className={style}
-        onClick={this.handleClick}
-      >
-        <span className={styles.inner}>
-          {
-            loading && <span className={styles.spinner}><Spinner inline /></span>
-          }
-          {
-            Icon && <span className={styles.iconContainer}><Icon className={styles.icon} /></span>
-          }
-          {
-            children && <span className={styles.content}>{children}</span>
-          }
-          {
-            ripple && !disabled && <Ink duration={200} opacity={0.10} radius={200} />
-          }
-        </span>
-      </a>
-    )
-  }
-}
+export default createButtonLike('a', {displayName: 'AnchorButton'})
